Split OfaTxContext implementations into multiple modules
