+++ crates/relayer-framework/src/base/one_for_all/impls/transaction/encode.rs
@@ -1,0 +1,24 @@
+use async_trait::async_trait;
+
+use crate::base::one_for_all::traits::transaction::OfaTxContext;
+use crate::base::one_for_all::types::transaction::OfaTxWrapper;
+use crate::base::transaction::traits::encode::CanEncodeTx;
+use crate::std_prelude::*;
+
+#[async_trait]
+impl<TxContext> CanEncodeTx for OfaTxWrapper<TxContext>
+where
+    TxContext: OfaTxContext,
+{
+    async fn encode_tx(
+        &self,
+        signer: &Self::Signer,
+        nonce: &Self::Nonce,
+        fee: &Self::Fee,
+        messages: &[Self::Message],
+    ) -> Result<Self::Transaction, Self::Error> {
+        self.tx_context
+            .encode_tx(signer, nonce, fee, messages)
+            .await
+    }
+}
