+++ crates/relayer-framework/src/base/one_for_all/impls/transaction/estimate.rs
@@ -1,0 +1,16 @@
+use async_trait::async_trait;
+
+use crate::base::one_for_all::traits::transaction::OfaTxContext;
+use crate::base::one_for_all::types::transaction::OfaTxWrapper;
+use crate::base::transaction::traits::estimate::CanEstimateTxFee;
+use crate::std_prelude::*;
+
+#[async_trait]
+impl<TxContext> CanEstimateTxFee for OfaTxWrapper<TxContext>
+where
+    TxContext: OfaTxContext,
+{
+    async fn estimate_tx_fee(&self, tx: &Self::Transaction) -> Result<Self::Fee, Self::Error> {
+        self.tx_context.estimate_tx_fee(tx).await
+    }
+}
